+++ history.json
@@ -10,5 +10,35 @@
     "username": "123",
     "date": "5/9/2024 3:31:1",
     "endpoint": "::1"
+  },
+  {
+    "message": "123",
+    "username": "123",
+    "date": "5/9/2024 3:58:16",
+    "endpoint": "::ffff:127.0.0.1"
+  },
+  {
+    "message": "123",
+    "username": "123",
+    "date": "5/9/2024 3:58:16",
+    "endpoint": "::ffff:127.0.0.1"
+  },
+  {
+    "message": "123",
+    "username": "123",
+    "date": "5/9/2024 3:58:17",
+    "endpoint": "::ffff:127.0.0.1"
+  },
+  {
+    "message": "123",
+    "username": "123",
+    "date": "5/9/2024 4:7:22",
+    "endpoint": "::1"
+  },
+  {
+    "message": "132",
+    "username": "123",
+    "date": "5/9/2024 4:7:23",
+    "endpoint": "::1"
   }
 ]
